Cleanup from refactor
